chore!: move GitRepo to the separate module
- GitError is now a derivative from CraftError

Closes: #385
