+++ craft_application/git/errors.py
@@ -1,0 +1,25 @@
+# Copyright 2024 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License version 3 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Git repository related errors."""
+
+from craft_cli.errors import CraftError
+
+
+class GitError(CraftError):
+    """Git is not working as expected."""
+
+    def __init__(self, details: str) -> None:
+        brief = "Git operation failed."
+        super().__init__(message=brief, details=details)
